+++ core/src/strategy/visualization.cpp
@@ -84,7 +84,7 @@ void Visualization::run(Robot& robot, Object& ball, Object& goal,
   // compute robot
   if (!config.serial.enabled) {
     // rotation
-    float max_rotation = robot.rotation_limit / config.strategy.fps / 3;
+    double max_rotation = robot.rotation_limit / config.strategy.fps / 3;
     robot.field_angle += clamp(robot.rotation, -1 * max_rotation, max_rotation);
     if (robot.emitter && override_ball) {
       ball.field_position = robot.position + robot_dir.resize(REAL_ROBOT_R);
